Added new historic data
The OECD had some historic tax rates for non-OECD countries (for the years 2000-2019) that had not yet been part of the TF dataset. They were now added. While all the datasets with the tax rates are updated, the calculations for the publications were not rerun, so they continue to match with the 2020 publication "Corporate Tax Rates around the World."